+++ app/Http/Controllers/v2/RsiaPksController.php
@@ -101,7 +101,7 @@ class RsiaPksController extends Controller
             }
         } catch (\Exception $e) {
             if ($request->hasFile('file')) {
-                if ($file && $st::disk('sftp')->exists(env('DOCUMENT_PKS_SAVE_LOCATION') . $fileName)) {
+                if ($file && $fileName != '' && $st::disk('sftp')->exists(env('DOCUMENT_PKS_SAVE_LOCATION') . $fileName)) {
                     \App\Helpers\Logger\RSIALogger::berkas("store data failed or file uploaded failed, deleting file", 'error', ['file_name' => $fileName, 'file_size' => $file->getSize(), 'data' => $request->all()]);
                     $st::disk('sftp')->delete(env('DOCUMENT_PKS_SAVE_LOCATION') . $fileName);
                 }
@@ -197,7 +197,7 @@ class RsiaPksController extends Controller
 
                     if ($request->hasFile('file')) {
                         $st = new \Illuminate\Support\Facades\Storage();
-                        if ($oldFIle && $st::disk('sftp')->exists(env('DOCUMENT_PKS_SAVE_LOCATION') . $oldFIle)) {
+                        if ($oldFIle && $oldFIle != '' && $st::disk('sftp')->exists(env('DOCUMENT_PKS_SAVE_LOCATION') . $oldFIle)) {
                             \App\Helpers\Logger\RSIALogger::berkas("file uploaded successfully | deleting old file", 'info', ['file_name' => $fileName, 'file_size' => $file->getSize(), 'data' => $request->all()]);
                             $st::disk('sftp')->delete(env('DOCUMENT_PKS_SAVE_LOCATION') . $oldFIle);
                         }
@@ -207,7 +207,7 @@ class RsiaPksController extends Controller
         } catch (\Exception $e) {
             if ($request->hasFile('file')) {
                 $st = new \Illuminate\Support\Facades\Storage();
-                if ($oldFIle && $st::disk('sftp')->exists(env('DOCUMENT_PKS_SAVE_LOCATION') . $oldFIle)) {
+                if ($oldFIle && $oldFIle != '' && $st::disk('sftp')->exists(env('DOCUMENT_PKS_SAVE_LOCATION') . $oldFIle)) {
                     \App\Helpers\Logger\RSIALogger::berkas("failed to update data or file uploaded failed, deleting file", 'error', ['file_name' => $fileName, 'file_size' => $file->getSize(), 'data' => $request->all()]);
                     $st::disk('sftp')->delete(env('DOCUMENT_PKS_SAVE_LOCATION') . $oldFIle);
                 }
@@ -245,7 +245,7 @@ class RsiaPksController extends Controller
         }
 
         $st = new \Illuminate\Support\Facades\Storage();
-        if ($data->berkas && $st::disk('sftp')->exists(env('DOCUMENT_PKS_SAVE_LOCATION') . $data->berkas)) {
+        if ($data->berkas && $data->berkas != '' && $st::disk('sftp')->exists(env('DOCUMENT_PKS_SAVE_LOCATION') . $data->berkas)) {
             \App\Helpers\Logger\RSIALogger::berkas("deleting file", 'info', ['file_name' => $data->berkas]);
             $st::disk('sftp')->delete(env('DOCUMENT_PKS_SAVE_LOCATION') . $data->berkas);
         }
